Changed repositories => Integer instead of Long; controller routes

+++ src/main/java/com/eEducation/ftn/repository/ClassRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.Class;
 
-public interface ClassRepository extends JpaRepository<Class, Long> {
+public interface ClassRepository extends JpaRepository<Class, Integer> {
 
 }
