+++ src/main/java/com/eEducation/ftn/repository/CourseLessonRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.CourseLesson;
 
-public interface CourseLessonRepository extends JpaRepository<CourseLesson, Long> {
+public interface CourseLessonRepository extends JpaRepository<CourseLesson, Integer> {
 
 }
