+++ src/main/java/com/eEducation/ftn/repository/StudentAttendsCourseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.StudentAttendsCourse;
 
-public interface StudentAttendsCourseRepository extends JpaRepository<StudentAttendsCourse, Long>{
+public interface StudentAttendsCourseRepository extends JpaRepository<StudentAttendsCourse, Integer>{
 
 }
