+++ src/main/java/com/eEducation/ftn/repository/StudentExamEntryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.StudentExamEntry;
 
-public interface StudentExamEntryRepository extends JpaRepository<StudentExamEntry, Long> {
+public interface StudentExamEntryRepository extends JpaRepository<StudentExamEntry, Integer> {
 
 }
