+++ src/main/java/com/eEducation/ftn/repository/TeacherTeachesCourseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eEducation.ftn.model.TeacherTeachesCourse;
 
-public interface TeacherTeachesCourseRepository extends JpaRepository<TeacherTeachesCourse, Long> {
+public interface TeacherTeachesCourseRepository extends JpaRepository<TeacherTeachesCourse, Integer> {
 
 }
